// src/erp-ui/projects/api/src/lib/model/custom-field-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ItemTypeBase } from './item-type-base';
import { CustomFieldType } from './custom-field-type';


export interface CustomFieldDto { 
    itemTypes: Array<ItemTypeBase>;
    id: string;
    name: string;
    label: string;
    type: CustomFieldType;
    required: boolean;
    order: number;
    selectOptions: Array<string>;
}