// src/erp-ui/projects/api/src/lib/model/item-type-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ItemTypeDto { 
    id: string;
    name: string;
    icon: string | null;
}